+++ src/app/layout.tsx
@@ -23,6 +23,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   // const initialState = cookieToInitialState(config, headers().get("cookie")) || null;
+  
   return (
     <html lang="en">
       <head>
