+++ src/app/page.tsx
@@ -1,5 +1,6 @@
 "use client";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
+import Image from "next/image";
 import { Button } from "@/components/ui/button";
 import {
   useAccount,
@@ -10,8 +11,12 @@ import {
 import { parseEther } from "viem";
 import { toast } from "sonner";
 import { useWeb3Modal } from "@web3modal/wagmi/react";
+import Claim from './claim'
+
+import {Markets, Forge, Swap, Virtualizer} from "./tabs"
 
 export default function Home() {
+  const [activeTab, setActiveTab] = useState<string>("");
   const { isConnected } = useAccount();
   const { signMessage } = useSignMessage();
   const { sendTransaction, data: hash } = useSendTransaction();
@@ -48,14 +53,67 @@ export default function Home() {
     }
   }, [isConfirming, isConfirmed, error, hash]);
 
+  const tabs = [
+    { id: 'Swap', label: 'Swap', logo: '/swap-logo.svg' },
+    { id: 'Virtualizer', label: 'Virtualizer', logo: "/virtualizer-logo.svg" },
+    { id: 'Forge', label: 'Forge', logo: "/forge-logo.svg" },
+    { id: 'Markets', label: 'Markets', logo: "/amm-logo.svg" },
+  ];
+
+
   return (
     <main>
-      <section className="py-12 flex flex-col items-center text-center gap-8">
-        <h1 className="text-4xl font-bold">Virtual Finance</h1>
-        <p className="text-2xl text-muted-foreground">
-          Welcome to our Testnet Dapp.
-        </p>
-      </section>
+      {!isConnected ? (
+          <>
+            <section className="py-12 flex flex-col items-center text-center gap-8">
+              <h1 className="text-4xl font-bold">Virtual Finance</h1>
+              <p className="text-2xl text-muted-foreground">
+                Welcome to our Testnet Dapp.
+              </p>
+            </section>
+            <Button onClick={handleConnect}>Connect Wallet</Button>
+          </>
+        ) : (
+          <>
+            <section className="py-12 flex flex-col items-center text-center gap-8">
+              <div className="flex gap-6 items-center justify-center">
+                <div className="font-medium text-center text-gray-500 border-b border-gray-200 dark:text-gray-400 dark:border-gray-700">
+                    {tabs.map((tab) => (
+                        <button
+                          key={tab.id}
+                          onClick={() => setActiveTab(tab.id)}
+                          className={`flex-1 py-2 px-4 text-center focus:outline-none ${
+                            activeTab === tab.id
+                              ? 'inline-block p-6 text-blue-600 border-b-2 border-blue-600 rounded-t-lg active dark:text-blue-500 dark:border-blue-500'
+                              : 'inline-block p-6 border-b-2 border-transparent rounded-t-lg hover:text-gray-600 hover:border-gray-300 dark:hover:text-gray-300'
+                          }`}
+                        >
+                          <>
+                          <Image
+                            src={tab.logo}
+                            alt="virtualizer-logo"
+                            width={24}
+                            height={24}
+                            priority
+                          />
+                          
+                          {tab.label}
+                          </>
+                        </button>
+                    ))}
+                  </div>
+              </div>
+            </section>
+              <section>
+                <div className="p-4">
+                  {activeTab === 'Swap' && <Swap />}
+                  {activeTab === 'Virtualizer' && <Virtualizer />}
+                  {activeTab === 'Forge' && <Forge />}
+                  {activeTab === 'Markets' && <Markets />}
+              </div>
+            </section>
+          </>
+      )}
     </main>
   );
 }
