+++ src/app/tabs/amm/Input.tsx
