+++ src/app/tabs/forge/EMC_to_VUSD.tsx
@@ -1,43 +1,38 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import InputComponent from "./Input";
-import { Button } from "../ui/button";
+import Image from "next/image";
+import { Button } from "../../../components/ui/button";
 import {
   useAccount,
-  useWriteContract,
   useWaitForTransactionReceipt,
+  useWriteContract,
 } from "wagmi";
 import { useWeb3Modal } from "@web3modal/wagmi/react";
-import VARQ_CONTRACT from "../../contracts/varq.json";
+import VARQ_CONTRACT from "../../../contracts/varq.json";
 import { Address, parseEther } from "viem";
 import { config } from "@/configs";
 import { getChainId } from "@wagmi/core";
 import { addresses } from "@/constants/addresses";
 import { toast } from "sonner";
 
-interface VUSDToEMCProps {
+interface EMCToVUSDProps {
   refreshBalance: () => void;
   balance: number;
 }
 
-const VUSD_to_EMC: React.FC<VUSDToEMCProps> = ({ refreshBalance, balance }) => {
+const EMC_to_VUSD: React.FC<EMCToVUSDProps> = ({ refreshBalance, balance }) => {
   const chainId = getChainId(config);
   const { address } = useAccount();
-  const { open } = useWeb3Modal();
-  const [destinationAddress, setDestinationAddress] = useState<Address>(
-    address || "0x"
-  );
   const handleConnect = () => {
     open();
   };
-
-  const [VUSD, setVUSD] = useState<number>(0);
-  const {
-    writeContract,
-    data: hash,
-    error: the_error,
-    isError,
-  } = useWriteContract();
-  const transfer_VUSD = parseEther(VUSD.toString());
+  const [destinationAddress, setDestinationAddress] = useState<Address>(
+    address || "0x"
+  );
+  const [VRT, setVRT] = useState<number>(0);
+  const { writeContract, data: hash } = useWriteContract();
+  const { open } = useWeb3Modal();
+  const transfer_VRT = String(parseEther(VRT.toString()));
 
   const handleDestinationAddress = (e: React.ChangeEvent<HTMLInputElement>) => {
     const newValue = e.target.value;
@@ -46,14 +41,15 @@ const VUSD_to_EMC: React.FC<VUSDToEMCProps> = ({ refreshBalance, balance }) => {
     );
   };
 
-  const handleVUSDtoEMC = () => {
+  const handleEMCtoVUSD = () => {
     writeContract({
       abi: VARQ_CONTRACT,
       address: addresses[chainId]["varq"],
-      functionName: "convertVUSDToTokens",
-      args: [transfer_VUSD, destinationAddress],
+      functionName: "convertTokensToVUSD",
+      args: [transfer_VRT, destinationAddress],
     });
-    console.log("Transferring:", VUSD + "to " + destinationAddress);
+
+    console.log("Transferring: ", [VRT] + "to " + [destinationAddress]);
   };
 
   const {
@@ -80,22 +76,23 @@ const VUSD_to_EMC: React.FC<VUSDToEMCProps> = ({ refreshBalance, balance }) => {
         },
       });
       refreshBalance?.();
-      setVUSD(0);
+      setVRT(0);
     }
     if (error) {
       toast.error("Transaction Failed");
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isConfirmed, isConfirming, error, hash]);
-
   return (
     <div>
       <div className="flex rounded-2xl items-left flex-col flex-grow pt-4 mx-2">
-        <h1 className="ml-2">vUSD -{">"} vTTD</h1>
+        <h1 className="ml-2">vTTD & vRT -{">"} vUSD</h1>
+      </div>
+      <div className="flex rounded-2xl items-left flex-col flex-grow mx-2">
         <InputComponent
-          label="vUSD"
-          onValueChange={setVUSD}
-          initialValue={VUSD}
+          label="vRT"
+          onValueChange={setVRT}
+          initialValue={VRT}
           balance={balance}
         />
         <p className="ml-2">Destination Address (Optional)</p>
@@ -110,7 +107,7 @@ const VUSD_to_EMC: React.FC<VUSDToEMCProps> = ({ refreshBalance, balance }) => {
       <div className="flex flex-col justify-center mx-2">
         {!address ? (
           <Button
-            className="bg-[#00A651] rounded-2xl px-6 hover:bg-[#C2D952]"
+            className="bg-[#00A651] rounded-2xl px-6 hover:bg-[#C2D952] font-semibold"
             onClick={handleConnect}
           >
             Connect Wallet
@@ -118,8 +115,8 @@ const VUSD_to_EMC: React.FC<VUSDToEMCProps> = ({ refreshBalance, balance }) => {
         ) : (
           <>
             <Button
-              className="bg-[#00A651] rounded-2xl px-6 hover:bg-[#C2D952]"
-              onClick={handleVUSDtoEMC}
+              className="bg-[#00A651] rounded-2xl px-6 hover:bg-[#C2D952] font-semibold"
+              onClick={handleEMCtoVUSD}
             >
               Convert
             </Button>
@@ -130,4 +127,4 @@ const VUSD_to_EMC: React.FC<VUSDToEMCProps> = ({ refreshBalance, balance }) => {
   );
 };
 
-export default VUSD_to_EMC;
+export default EMC_to_VUSD;
