+++ src/app/tabs/index.ts
@@ -1,0 +1,6 @@
+import Markets from "./amm"
+import Forge from "./forge"
+import Swap from "./swap"
+import Virtualizer from "./virtualizer"
+
+export {Markets, Forge, Swap, Virtualizer}
