+++ src/app/tabs/swap/VexModal.tsx
@@ -1,7 +1,8 @@
 import React, { useEffect } from "react";
-import { Button } from "../ui/button";
+import { Button } from "@/components/ui/button";
 import { useWriteContract, useWaitForTransactionReceipt } from "wagmi";
-import MUSD_CONTRACT from "../../contracts/mUSD.json";
+import VRT_CONTRACT from "../../../contracts/vtoken.json";
+import VEX_CONTRACT from "@/contracts/vex.json";
 import { config } from "@/configs";
 import { getChainId } from "@wagmi/core";
 import { addresses } from "@/constants/addresses";
@@ -9,16 +10,18 @@ import { toast } from "sonner";
 
 interface ModalProps {
   isOpen: boolean;
-  onClose: () => void; // Define the type for onClose as a function that returns nothing
+  onClose: () => void;
   children: React.ReactNode;
-  refetchApproval: () => void;
+  swapType: string;
+  refetchApprovals: () => void;
 }
 
 const Modal: React.FC<ModalProps> = ({
   isOpen,
   onClose,
   children,
-  refetchApproval,
+  swapType,
+  refetchApprovals,
 }) => {
   const chainId = getChainId(config);
   const {
@@ -27,15 +30,6 @@ const Modal: React.FC<ModalProps> = ({
     isError: error,
     data: hash,
   } = useWriteContract();
-
-  useEffect(() => {
-    if (isSuccess === true) {
-      onClose();
-    }
-    if (error) {
-      console.error("Transaction Failed");
-    }
-  }, [isSuccess, error, onClose]);
 
   const {
     isLoading: isConfirming,
@@ -60,7 +54,8 @@ const Modal: React.FC<ModalProps> = ({
           },
         },
       });
-      refetchApproval?.();
+      onClose();
+      refetchApprovals?.();
     }
     if (error) {
       toast.error("Approval Failed");
@@ -70,32 +65,39 @@ const Modal: React.FC<ModalProps> = ({
 
   if (!isOpen) return null;
 
+  const handleApprove = () => {
+    const abi = VRT_CONTRACT;
+    const address =
+      swapType === "vUSD"
+        ? addresses[chainId]["vusd"]
+        : addresses[chainId]["vttd"]; // Replace with actual VTTD contract address
+
+    writeContractAsync({
+      abi,
+      address,
+      functionName: "approve",
+      args: [addresses[chainId]["vex"], "10000000000000000000000000"],
+    });
+  };
+
   return (
     <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center p-4">
-      <div className="bg-white dark:bg-[#1e293b] rounded-lg shadow-2xl p-6 w-full max-w-md">
+      <div className="bg-secondary rounded-lg shadow-lg p-6 w-full max-w-md">
         {children}
         <div className="flex flex-row justify-between">
           <Button
             onClick={onClose}
-            className="mt-4 py-2 px-4 bg-[#020817] bg-[#F15A22] hover:bg-[#F5846F]"
+            disabled={isConfirming}
+            className="mt-4 py-2 px-4 bg-[#020817] text-white"
           >
             Cancel
           </Button>
           <Button
-            onClick={() =>
-              writeContractAsync({
-                abi: MUSD_CONTRACT,
-                address: addresses[chainId]["musd"],
-                functionName: "approve",
-                args: [
-                  addresses[chainId]["virtualizer"],
-                  "1000000000000000000000000",
-                ],
-              })
-            }
-            className="mt-4 py-2 px-4 bg-[#F15A22] hover:bg-[#F5846F]"
+            onClick={handleApprove}
+            disabled={isConfirming}
+            className="mt-4 py-2 px-4 bg-[#020817] text-white"
           >
-            Approve
+            Approve {swapType.toUpperCase()}
           </Button>
         </div>
       </div>
