+++ src/app/tabs/virtualizer/WithdrawWidget.tsx
@@ -1,35 +1,50 @@
 import React, { useState, useEffect } from "react";
-import { Button } from "../ui/button";
-import InputComponent from "./Input";
+import { Button } from "../../../components/ui/button";
+import InputComponent from "@/app/tabs/virtualizer/Input";
 import Modal from "./Modal";
 import DisabledInputComponent from "./DisabledInput";
+import VIRTUALISER_CONTRACT from "../../../contracts/virtualizer.json";
 import {
   useAccount,
   useReadContract,
-  useWaitForTransactionReceipt,
   useWriteContract,
+  useWaitForTransactionReceipt,
 } from "wagmi";
-import MUSD_CONTRACT from "../../contracts/mUSD.json";
-import VIRTUALISER_CONTRACT from "../../contracts/virtualizer.json";
 import { useWeb3Modal } from "@web3modal/wagmi/react";
+import MUSD_CONTRACT from "../../../contracts/mUSD.json";
 import { parseEther } from "viem";
 import { config } from "@/configs";
 import { getChainId } from "@wagmi/core";
 import { addresses } from "@/constants/addresses";
 import { toast } from "sonner";
 
-interface DepositWidgetProps {
+interface CustomToastProps {
+  message: string;
+  gifUrl: string;
+}
+
+const CustomToast: React.FC<CustomToastProps> = ({ message, gifUrl }) => (
+  <div className="flex flex-col items-center">
+    <img
+      src={gifUrl}
+      alt="Toast Icon"
+      className="border border-green-400 self-center"
+    />
+    <h1 className="text-xl font-bold">{message}...</h1>
+  </div>
+);
+interface WithdrawWidgetProps {
   refreshBalance: () => void; // Define the type of refreshBalance as a function
   balance: number | null;
 }
 
-const DepositWidget: React.FC<DepositWidgetProps> = ({
+const WithdrawWidget: React.FC<WithdrawWidgetProps> = ({
   refreshBalance,
   balance,
 }) => {
   const chainId = getChainId(config);
   const { address } = useAccount();
-  const [mUSDC, setmUSDC] = useState<number>(0);
+  const [vUSD, setvUSD] = useState<number>(0);
   const [isModalOpen, setModalOpen] = useState(false);
   const { writeContract, data: hash } = useWriteContract();
   const openModal = () => setModalOpen(true);
@@ -55,26 +70,43 @@ const DepositWidget: React.FC<DepositWidgetProps> = ({
   });
 
   const approve_str = approved?.toString();
-  const transfer_mUSDC = String(parseEther(mUSDC.toString()));
+  const transfer_vUSD = String(parseEther(vUSD.toString()));
 
-  const handleDeposit = () => {
+  const handleWithdraw = () => {
     if (approve_str === "0") {
       openModal();
     } else {
       writeContract({
         abi: VIRTUALISER_CONTRACT,
         address: addresses[chainId]["virtualizer"],
-        functionName: "wrap",
-        args: [transfer_mUSDC],
+        functionName: "unwrap",
+        args: [transfer_vUSD],
       });
 
-      console.log("Transferring:", transfer_mUSDC);
+      console.log("Transferring:", transfer_vUSD);
     }
   };
 
   useEffect(() => {
     if (isConfirming) {
-      toast.loading("Transaction Pending");
+      toast.loading(
+        <CustomToast
+          message="TransactionPending"
+          gifUrl="walking_orange.gif"
+        />,
+        {
+          style: {
+            background: "#3A4047",
+            width: "33vw",
+            height: "75vh",
+            top: "50%",
+            left: "50%",
+            transform: "translate(-50%,-50%)",
+            position: "fixed",
+          },
+          //className:"bg-[#3A4047] w-full h-full top-[145px] left-[490px]"
+        }
+      );
     }
     toast.dismiss();
 
@@ -88,7 +120,7 @@ const DepositWidget: React.FC<DepositWidgetProps> = ({
         },
       });
       refreshBalance?.();
-      setmUSDC(0);
+      setvUSD(0);
     }
     if (error) {
       toast.error("Transaction Failed");
@@ -102,10 +134,10 @@ const DepositWidget: React.FC<DepositWidgetProps> = ({
         <div>
           <div className="flex mx-2">
             <InputComponent
-              label="mUSDC"
-              type="deposit"
-              initialValue={mUSDC}
-              onValueChange={setmUSDC}
+              label="vUSD"
+              type="withdraw"
+              initialValue={vUSD}
+              onValueChange={setvUSD}
               balance={balance}
             />
           </div>
@@ -123,10 +155,10 @@ const DepositWidget: React.FC<DepositWidgetProps> = ({
         ) : (
           <Button
             className="w-full rounded-full bg-[#F15A22] hover:bg-[#F5846F] font-semibold"
-            onClick={handleDeposit}
+            onClick={handleWithdraw}
             disabled={isConfirming}
           >
-            Deposit
+            Withdraw
           </Button>
         )}
 
@@ -136,10 +168,10 @@ const DepositWidget: React.FC<DepositWidgetProps> = ({
           refetchApproval={refetch_approval}
         >
           <DisabledInputComponent
-            label="mUSDC"
+            label="vUSD"
             heading="Approve amount of funds that can be transferred"
             initialValue={1000000}
-            currency="mUSDC"
+            currency="vUSD"
           />
         </Modal>
       </div>
@@ -147,4 +179,4 @@ const DepositWidget: React.FC<DepositWidgetProps> = ({
   );
 };
 
-export default DepositWidget;
+export default WithdrawWidget;
