+++ src/app/tabs/virtualizer/index.tsx
@@ -3,11 +3,11 @@ import { Card } from "@tremor/react";
 import Image from "next/image";
 import React, { useEffect, useState } from "react";
 import CustomTab from "@/components/CustomTab";
-import DepositWidget from "@/components/virtualizer/DepositWidget";
-import WithdrawWidget from "@/components/virtualizer/WithdrawWidget";
+import DepositWidget from "./DepositWidget";
+import WithdrawWidget from "@/app/tabs/virtualizer/WithdrawWidget";
 import { useAccount, useReadContract } from "wagmi";
-import MUSD_CONTRACT from "../../contracts/mUSD.json";
-import VUSD_CONTRACT from "../../contracts/vtoken.json";
+import MUSD_CONTRACT from "@/contracts/mUSD.json";
+import VUSD_CONTRACT from "@/contracts/vtoken.json";
 import { config } from "@/configs";
 import { getChainId } from "@wagmi/core";
 import { addresses } from "@/constants/addresses";
