+++ src/components/Nav.tsx
@@ -3,26 +3,16 @@ import Image from "next/image";
 import { ModeToggle } from "./Modetoggle";
 import Link from "next/link";
 import { useEffect, useState } from "react";
-import { usePathname } from "next/navigation";
-
 
 export default function Nav() {
-  const pathname = usePathname();
   const [activeTab, setActiveTab] = useState<string>("");
-
-  useEffect(() => {
-    const tabName = pathname.split('/')[1];
-    setActiveTab(tabName);
-    localStorage.setItem('activeTab', tabName);
-  }, [pathname]);
-  
 
   useEffect(() => {
     const storedActiveTab = localStorage.getItem("activeTab");
     if (storedActiveTab) {
       setActiveTab(storedActiveTab);
     }
-  }, [activeTab]);
+  }, []);
 
   const handleTabChange = (tabName: string) => {
     setActiveTab(tabName);
@@ -50,114 +40,7 @@ export default function Nav() {
                   priority
                 />
               </a>
-            </li>
-            <Link
-              className={`p-2 rounded-2xl hover:bg-secondary ${
-                activeTab === "claim"
-                  ? "bg-secondary text-white"
-                  : "text-gray-500"
-              }`}
-              href="claim"
-            >
-              Claim
-            </Link>
-
-            <Link href="swap" onClick={() => handleTabChange("swap")}>
-              <div
-                className={`flex flex-row group hover:text-[#ffcb03] hover:${
-                  activeTab === "swap"
-                } ${activeTab === "swap" ? "text-[#ffcb03]" : "text-gray-500"}`}
-              >
-                <Image
-                  src="/swap-logo.svg"
-                  alt="swap-logo"
-                  width={24}
-                  height={24}
-                  priority
-                  className={`mr-1 ${
-                    activeTab === "swap"
-                      ? ""
-                      : "grayscale group-hover:grayscale-0"
-                  }`}
-                />
-                Swap
-              </div>
-            </Link>
-
-            <Link
-              href="virtualizer"
-              onClick={() => handleTabChange("virtualizer")}
-            >
-              <div
-                className={`flex flex-row group hover:text-[#F15A22] hover:${
-                  activeTab === "virtualizer"
-                } ${
-                  activeTab === "virtualizer"
-                    ? "text-[#F15A22]"
-                    : "text-gray-500"
-                }`}
-              >
-                <Image
-                  src="/virtualizer-logo.svg"
-                  alt="virtualizer-logo"
-                  width={24}
-                  height={24}
-                  priority
-                  className={`mr-1 ${
-                    activeTab === "virtualizer"
-                      ? ""
-                      : "grayscale group-hover:grayscale-0"
-                  }`}
-                />
-                Virtualizer
-              </div>
-            </Link>
-
-            <Link href="forge" onClick={() => handleTabChange("forge")}>
-              <div
-                className={`flex flex-row group hover:text-[#00A651] hover:${
-                  activeTab === "forge"
-                } ${
-                  activeTab === "forge" ? "text-[#00A651]" : "text-gray-500"
-                }`}
-              >
-                <Image
-                  src="/forge-logo.svg"
-                  alt="forge-logo"
-                  width={24}
-                  height={24}
-                  priority
-                  className={`mr-1 ${
-                    activeTab === "forge"
-                      ? ""
-                      : "grayscale group-hover:grayscale-0"
-                  }`}
-                />
-                Forge
-              </div>
-            </Link>
-
-            <Link href="amm" onClick={() => handleTabChange("amm")}>
-              <div
-                className={`flex flex-row group hover:text-[#68AAFF] hover:${
-                  activeTab === "amm"
-                } ${activeTab === "amm" ? "text-[#68AAFF]" : "text-gray-500"}`}
-              >
-                <Image
-                  src="/amm-logo.svg"
-                  alt="amm-logo"
-                  width={24}
-                  height={24}
-                  priority
-                  className={`mr-1 ${
-                    activeTab === "amm"
-                      ? ""
-                      : "grayscale group-hover:grayscale-0"
-                  }`}
-                />
-                AMM
-              </div>
-            </Link>
+            </li> 
           </ul>
           <div className="flex flex-row items-center">
             <div className="mr-2">
